pyproject.toml
[project]
name = "api-key-service"
version = "0.2.0"
dependencies = [
  "fastapi",
  "uvicorn",
  "sqlalchemy",
  "psycopg2-binary",
  "pydantic-settings",
  "passlib[bcrypt]",
  "PyJWT",
  "bcrypt==4.0.1"
]